Refactor: Update calculation of weighted sum in dequad function
The dequad function in dequad.py has been modified to update the calculation of the weighted sum. Instead of multiplying the weights with the function values separately, the weights are now multiplied with the entire array of function values. This change improves the efficiency of the calculation.

Refactor: Improve efficiency of sigmalos2_dequad function

The sigmalos2_dequad function in model.py has been optimized for better performance. The unnecessary line of code that was previously present has been removed, resulting in improved efficiency.